requirements.txt
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.25.0
python-docx>=1.1.0
Pillow>=9.0.0
customtkinter